pyproject.toml
[project]
name = "medical-coding-env"
version = "0.1.0"
description = "Medical Coding OpenEnv Environment"
requires-python = ">=3.10"

dependencies = [
    "openenv-core",
    "fastapi",
    "uvicorn",
    "pydantic",
    "openai"
]

[tool.setuptools]
packages = ["med_env"]

[project.scripts]
server = "server.app:main"